[v0.91.5][rust-refactor][3/6] Refactor prompt-template editor domain boundaries (Closes #3748)
